Use secure location for facter_dot_d cache
Creating files with predictable names in /tmp is a well-known security
exploit. It would probably be better to base the default on a
configuration parameter that is settable at runtime, but I do not have
time for that and anyway this functionality has been merged into Facter
itself, so a more thorough fix would be a dead end. (Facter does not
seem to have the same problem, AFAICT.)